User authentication added

+++ controllers/authorsController.js
@@ -1,18 +1,12 @@
-const { logosObj } = require("../exports/util");
 const { internalErrorRes } = require("../exports/helpers");
 const { Authors } = require("../exports/models");
 const crypto = require("crypto");
 
 exports.getAuthors = async (req, res, next) => {
   try {
-    const authorsRes = await Authors.findAll();
+    const authorsRes = await Authors.findAll({ where: { user_id: req.session.user.id } });
     const authors = authorsRes.map((res) => res.dataValues);
-
-    res.render("authors/admin-authors", {
-      logosObj,
-      authors,
-      noAuthors: authors.length === 0,
-    });
+    res.render("authors/admin-authors", { authors });
   } catch (error) {
     console.log(`\nError: ${error}\n`);
     internalErrorRes(res);
@@ -20,23 +14,19 @@ exports.getAuthors = async (req, res, next) => {
 };
 
 exports.getAddAuthor = (req, res, next) => {
-  res.render("authors/save-author", { logosObj, edit: false });
+  res.render("authors/save-author", { edit: false });
 };
 
 exports.getEditAuthor = async (req, res, next) => {
   try {
     const id = req.params.id;
-    const authorsRes = await Authors.findOne({ where: { id } });
+    const authorsRes = await Authors.findOne({ where: { id, user_id: req.session.user.id } });
 
     if (!authorsRes) return res.redirect("/admin-authors");
 
     const author = authorsRes.dataValues;
 
-    res.render("authors/save-author", {
-      logosObj,
-      edit: true,
-      author,
-    });
+    res.render("authors/save-author", { edit: true, author });
   } catch (error) {
     console.log(`\nError: ${error}\n`);
     internalErrorRes(res);
@@ -46,15 +36,14 @@ exports.getEditAuthor = async (req, res, next) => {
 exports.getDeleteAuthor = async (req, res, next) => {
   try {
     const id = req.params.id;
-    const authorsRes = await Authors.findOne({ where: { id } });
+    const authorsRes = await Authors.findOne({ where: { id, user_id: req.session.user.id } });
 
     if (!authorsRes) return res.redirect("/admin-authors");
 
-    const { name, related_books } = authorsRes.dataValues;
+    const { name } = authorsRes.dataValues;
 
     res.render("confirm", {
-      logosObj,
-      cascade: related_books > 0,
+      isNotBook: true,
       page: "authors",
       model: "Autor",
       modelMsg: "este autor",
@@ -74,12 +63,13 @@ exports.postAddAuthor = async (req, res, next) => {
     if (name && email) {
       await Authors.create({
         id: crypto.randomUUID(),
+        user_id: req.session.user.id,
         name,
-        email,
-        related_books: 0,
+        email
       });
     }
 
+    req.flash("msg", "Autor creado exitosamente");
     res.redirect("/admin-authors");
   } catch (error) {
     console.log(`\nError: ${error}\n`);
@@ -92,9 +82,13 @@ exports.postEditAuthor = async (req, res, next) => {
     const { id, name, email } = req.body;
 
     if (id && name && email) {
-      await Authors.update({ name, email }, { where: { id } });
+      await Authors.update(
+        { name, email },
+        { where: { id, user_id: req.session.user.id } }
+      );
     }
 
+    req.flash("msg", "Autor editado exitosamente");
     res.redirect("/admin-authors");
   } catch (error) {
     console.log(`\nError: ${error}\n`);
@@ -105,7 +99,8 @@ exports.postEditAuthor = async (req, res, next) => {
 exports.postDeleteAuthor = async (req, res, next) => {
   try {
     const id = req.body.id;
-    await Authors.destroy({ where: { id } });
+    await Authors.destroy({ where: { id, user_id: req.session.user.id } });
+    req.flash("msg", "Autor eliminado exitosamente");
     res.redirect("/admin-authors");
   } catch (error) {
     console.log(`\nError: ${error}\n`);
