+++ controllers/editorialsController.js
@@ -1,19 +1,12 @@
 const { Editorials } = require("../exports/models");
-const { logosObj } = require("../exports/util");
 const { internalErrorRes } = require("../exports/helpers");
-const sequelize = require("sequelize");
 const crypto = require("crypto");
 
 exports.getEditorials = async (req, res, next) => {
   try {
-    const EditorialsRes = await Editorials.findAll();
+    const EditorialsRes = await Editorials.findAll({ where: { user_id: req.session.user.id } });
     const editorials = EditorialsRes.map((res) => res.dataValues);
-
-    res.render("editorials/admin-editorials", {
-      logosObj,
-      editorials,
-      noEditorials: editorials.length === 0,
-    });
+    res.render("editorials/admin-editorials", { editorials });
   } catch (error) {
     console.log(`\nError: ${error}\n`);
     internalErrorRes(res);
@@ -21,23 +14,21 @@ exports.getEditorials = async (req, res, next) => {
 };
 
 exports.getAddEditorial = (req, res, next) => {
-  res.render("editorials/save-editorial", { logosObj, edit: false });
+  res.render("editorials/save-editorial", { edit: false });
 };
 
 exports.getEditEditorial = async (req, res, next) => {
   try {
     const id = req.params.id;
-    const editorialsRes = await Editorials.findOne({ where: { id } });
+    const editorialsRes = await Editorials.findOne({
+      where: { id, user_id: req.session.user.id }
+    });
 
     if (!editorialsRes) return res.redirect("/admin-editorials");
 
     const editorial = editorialsRes.dataValues;
 
-    res.render("editorials/save-editorial", {
-      logosObj,
-      edit: true,
-      editorial,
-    });
+    res.render("editorials/save-editorial", { edit: true, editorial });
   } catch (error) {
     console.log(`\nError: ${error}\n`);
     internalErrorRes(res);
@@ -47,15 +38,16 @@ exports.getEditEditorial = async (req, res, next) => {
 exports.getDeleteEditorial = async (req, res, next) => {
   try {
     const id = req.params.id;
-    const editorialsRes = await Editorials.findOne({ where: { id } });
+    const editorialsRes = await Editorials.findOne({
+      where: { id, user_id: req.session.user.id }
+    });
 
     if (!editorialsRes) return res.redirect("/admin-editorials");
 
-    const { name, related_books } = editorialsRes.dataValues;
+    const { name } = editorialsRes.dataValues;
 
     res.render("confirm", {
-      logosObj,
-      cascade: related_books > 0,
+      isNotBook: true,
       page: "editorials",
       model: "Editorial",
       modelMsg: "esta editorial",
@@ -75,13 +67,14 @@ exports.postAddEditorial = async (req, res, next) => {
     if (phone && name && country) {
       await Editorials.create({
         id: crypto.randomUUID(),
+        user_id: req.session.user.id,
         name,
         phone,
-        country,
-        related_books: 0,
+        country
       });
     }
 
+    req.flash("msg", "Editorial creada exitosamente");
     res.redirect("/admin-editorials");
   } catch (error) {
     console.log(`\nError: ${error}\n`);
@@ -94,9 +87,13 @@ exports.postEditEditorial = async (req, res, next) => {
     const { name, phone, country, id } = req.body;
 
     if (name && phone && country && id) {
-      await Editorials.update({ name, phone, country }, { where: { id } });
+      await Editorials.update(
+        { name, phone, country },
+        { where: { id, user_id: req.session.user.id } }
+      );
     }
 
+    req.flash("msg", "Editorial editada exitosamente");
     res.redirect("/admin-editorials");
   } catch (error) {
     console.log(`\nError: ${error}\n`);
@@ -107,7 +104,8 @@ exports.postEditEditorial = async (req, res, next) => {
 exports.postDeleteEditorial = async (req, res, next) => {
   try {
     const id = req.body.id;
-    await Editorials.destroy({ where: { id } });
+    await Editorials.destroy({ where: { id, user_id: req.session.user.id } });
+    req.flash("msg", "Editorial eliminada exitosamente");
     res.redirect("/admin-editorials");
   } catch (error) {
     console.log(`\nError: ${error}\n`);
